Successfully smuggle two Completer instances through built_redux
With some effort, I have successfully managed to smuggle a new class called SpecializedCompleterTuple through some actions such that the middleware can have some influence on a UI widget which originally constructed (and is listening to the futures inside of) the SpecializedCompleterTuple. My purposes for doing this are outlined in the documentation of SpecializedCompleterTuple.

I'm justifying how jank this system is because it's not far off from what Brian Egan recommended (https://github.com/brianegan/flutter_redux/issues/6#issuecomment-365720043). However, I do hope I find something cleaner than this for the future. This also relates to davidmarne/built_redux#103 and flutter/flutter#11675.