Isolated business logic and client components
I've also updated the behavior of the build process in .travis.yml as compilation and testing has changed (slightly).

This commit comes just before a large migration towards async_redux which makes dealing with redux a lot less of a hassle. The reason for isolating business logic and client components is per a recommendation (that I happen to agree with) from the package author. See https://pub.dev/packages/async_redux#recommended-directory-structure